Update requirements, supported package versions, tox
Many of the requirement files used for testing different package
versions in different python environments using tox were not internally
consistent on all platforms/in all combinations. This commit is the
first step to fixing that inconsistency.

Also I removed tox and pytest from all requirement.txt files as they are
only needed to test the package, not to just run it.

In concord with the above, the documentation now recommends using
requirements.txt, instead of the old environment_rescomp.yml to create
the conda environment.